Refactor createProduct and listProducts routes to use separate directories for each route
